TC/requirements.txt
pytest
httpx
PyJWT
python-dotenv
pytest-asyncio
psycopg2-binary